 requirements.txt
kivy
google-cloud-vision
